# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Syon API
  description: API REST para programação e cybersegurança
  version: 1.0.0b0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
...
  /v1/chat/completions:
    post:
      summary: Create chat completion
      tags: [chat]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat completion generated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ChatRequest:
      type: object
      required: [messages]
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
...